test-runner: initialize gcc-torture submodule + clean up unused vars
- git submodule update --init z80-utils/vendor/gcc-torture (torture suite
  now runs: 3308 PASS / 4 XFAIL / 3 FAIL — all known clang 16-bit limits)
- Remove dead rt_dir/rt_name locals in llc.rs (were left over after the
  direct-lib fix, now causing compiler warnings)
- Fix test_96 memset signature: use unsigned int for size param

+++ z80-utils/test-runner/testcases/clang/test_96_ldir_optimizations.c
@@ -11,7 +11,7 @@ typedef unsigned short uint16_t;
 
 extern void *memcpy(void *dest, const void *src, uint16_t n);
 extern void *memmove(void *dest, const void *src, uint16_t n);
-extern void *memset(void *p, int v, uint16_t n);
+extern void *memset(void *p, int v, unsigned int n);
 
 static uint8_t src_buf[32];
 static uint8_t dst_buf[32];
